Move `apply_content_changes()` to own file
